+++ apps/antalmanac/src/components/Calendar/Toolbar/ScheduleSelect/drag-and-drop/SortableItem.tsx
@@ -1,11 +1,9 @@
 import type { DraggableSyntheticListeners, UniqueIdentifier } from '@dnd-kit/core';
 import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
-import DragIndicatorIcon from '@mui/icons-material/DragIndicator';
-import { Box, ListItem } from '@mui/material';
-import { useTheme } from '@mui/material/styles';
+import { ListItem } from '@mui/material';
 import type { CSSProperties, PropsWithChildren } from 'react';
-import { createContext, useContext, useMemo } from 'react';
+import { createContext, useMemo } from 'react';
 
 interface Props {
     id: UniqueIdentifier;
@@ -17,7 +15,7 @@ interface Context {
     ref(node: HTMLElement | null): void;
 }
 
-const SortableItemContext = createContext<Context>({
+export const SortableItemContext = createContext<Context>({
     attributes: {},
     listeners: undefined,
     ref: (_node: HTMLElement | null) => {
@@ -49,41 +47,5 @@ export function SortableItem({ children, id }: PropsWithChildren<Props>) {
                 {children}
             </ListItem>
         </SortableItemContext.Provider>
-    );
-}
-
-interface DragHandleProps {
-    disabled?: boolean;
-}
-
-export function DragHandle({ disabled = false }: DragHandleProps) {
-    const { attributes, listeners, ref } = useContext(SortableItemContext);
-    const theme = useTheme();
-
-    return (
-        <Box
-            {...attributes}
-            {...(disabled ? {} : listeners)}
-            ref={ref}
-            sx={{
-                display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'center',
-                cursor: disabled ? 'auto' : 'pointer',
-                borderRadius: 1,
-                '&:hover': {
-                    backgroundColor: disabled ? 'transparent' : 'rgba(0, 0, 0, 0.1)',
-                },
-                '&:focus-visible': {
-                    boxShadow: disabled ? 'none' : '0 0 0 2px #4c9ffe',
-                },
-            }}
-        >
-            <DragIndicatorIcon
-                style={{
-                    color: disabled ? 'gray' : theme.palette.mode === 'light' ? 'black' : 'white',
-                }}
-            />
-        </Box>
     );
 }
